pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multimodal-agent"
version = "0.1.0"
description = "Multi-modal AI agent for images, text, and documents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pillow>=10.4.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.24", "ruff>=0.7.0"]
